% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_cols}
\alias{add_cols}
\title{Add Missing Columns
This function adds missing columns to one data frame that are}
\usage{
add_cols(df1, df2)
}
\arguments{
\item{df1}{}

\item{df2}{}
}
\description{
Add Missing Columns
This function adds missing columns to one data frame that are
}
