% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clip_sf.R
\name{clip_sf}
\alias{clip_sf}
\title{Clip sf object to polygon}
\usage{
clip_sf(sf_lyr, sf_clip, locale)
}
\arguments{
\item{sf_lyr}{sf object. Spatial data to be clipped.}

\item{sf_clip}{sf polygon object. Spatial layer to clip data by.}

\item{locale}{Character. Short description of clipped layer, usually the
location (e.g., Forest acronym or "Buffer").}
}
\value{
sf object
}
\description{
This function clips a sf object using \code{sf::st_intersection()}. First, this
function checks that the coordinate reference system (CRS) of the input
object is the same as the clipping object. If it is not, this function
transforms the CRS of the input object to the clipping object using
\code{sf::st_transform()}before clipping. The returned sf object will be in
the clipping object CRS if they are not the same.
}
\examples{
## Not run:
# Clip species occurrence point data to the administrative boundary of a 
# forest.
clip_sf(sf_lyr = spocc_sf, sf_clip = admin_bdy_sf, locale = "USFS")

## End (Not run)
}
