% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_dir}
\alias{create_dir}
\title{Create project directories
This function evaluates if a project directory exists. If one does not, it
creates a project directory. This function also makes a copy of the
species data pull report (*.Rmd) in the project directory.}
\usage{
create_dir(unit_dir)
}
\arguments{
\item{unit_dir}{Character. File path of the project directory.}
}
\description{
Create project directories
This function evaluates if a project directory exists. If one does not, it
creates a project directory. This function also makes a copy of the
species data pull report (*.Rmd) in the project directory.
}
