% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_taxonomies.R
\name{get_taxonomies}
\alias{get_taxonomies}
\title{adds taxonomy information from GBIF to any df with scientific names and returns a tibble. Returns gbif_taxonID which is the GBIF ID for the given scientific name and full taxonomy for the accepted taxonomy from the GBIF backbone taxonomies database. taxon_id as the accepted taxonomy ID from GBIF backbone.}
\usage{
get_taxonomies(x, query_field = "scientific_name")
}
\arguments{
\item{x}{a dataframe containing at least on column including scientific name in species or subspecies form.}

\item{query_field}{the name of the field with scientific names. Defaults to "scientific_name"}
}
\description{
adds taxonomy information from GBIF to any df with scientific names and returns a tibble. Returns gbif_taxonID which is the GBIF ID for the given scientific name and full taxonomy for the accepted taxonomy from the GBIF backbone taxonomies database. taxon_id as the accepted taxonomy ID from GBIF backbone.
}
