% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{id_vec}
\alias{id_vec}
\title{Create vector from comma separated variable (column)
This function creates a vector from a comma separated variable or column,
then removes NAs and blanks}
\usage{
id_vec(id_var)
}
\arguments{
\item{id_var}{Input variable, usually df$var (e.g., spp_list$GBIF_taxonKey)}
}
\description{
Create vector from comma separated variable (column)
This function creates a vector from a comma separated variable or column,
then removes NAs and blanks
}
