% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ns_ranks.R
\name{ns_ranks}
\alias{ns_ranks}
\title{Generate a dataframe of species names with NatureServe G- and S-ranks}
\usage{
ns_ranks(species_list, states = NULL)
}
\arguments{
\item{species_list}{A character vector of species names.}

\item{states}{A character vector of 2-letter US state codes. Default is NULL, if
NULL all US states are returned.}
}
\value{
tibble::tibble()
}
\description{
This function uses the \code{natserv} package to query the NatureServe API
and read those data into R. This function then convers the JSON
data returned into a \code{tibble::tibble()}.
}
\examples{
spp = c("Buteo jamaicensis", "Falco peregrinus", "Aquila chrysaetos")
ns_ranks(spp, c("UT", "CO"))
}
