% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fc.R
\name{read_fc}
\alias{read_fc}
\title{Read feature class into R.
This function uses the \code{sf} package to read a feature class into R from a
geodatabase. It then checks that the sf object is valid and validates the
object if it is not valid.}
\usage{
read_fc(lyr, dsn)
}
\arguments{
\item{lyr}{Character. Feature class name.}

\item{dsn}{Character. Path to geodatabase that holds the feature class.}
}
\value{
sf objectread_fc <- function(lyr, dsn, crs){
}
\description{
Read feature class into R.
This function uses the \code{sf} package to read a feature class into R from a
geodatabase. It then checks that the sf object is valid and validates the
object if it is not valid.
}
\examples{
## Not run:
# Read a feature class into R
read_lyr(lyr = "admin_bdy", dsn = file.path("T:/path/to/geodatabase"))

## End (Not run)
}
