% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_list_ex.R
\docType{data}
\name{sp_list_ex}
\alias{sp_list_ex}
\title{An example dataset to test functions that consists of a list of species with common_name and scientific_name.}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 427 rows and 2 columns.
}
\usage{
sp_list_ex
}
\description{
An example dataset to test functions that consists of a list of species with common_name and scientific_name.
}
\keyword{datasets}
