% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{view_dups}
\alias{view_dups}
\title{View duplicate species in a data frame}
\usage{
view_dups(spp_dat, spp_vec)
}
\arguments{
\item{spp_dat}{Data frame or tibble of species data.}

\item{spp_vec}{Vector of species names from the same data frame.}
}
\description{
View duplicate species in a data frame
}
