% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{wkt_string}
\alias{wkt_string}
\title{Create a WTK string
Creates a well-known text string from a polygon (sf object).}
\usage{
wkt_string(my_polygon)
}
\arguments{
\item{my_polygon}{sf object. An sf polygon object.}
}
\description{
Create a WTK string
Creates a well-known text string from a polygon (sf object).
}
